+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Features
 
 - Pre-main profiling data is now attached to app start spans (#3736)
+## 8.22.3-beta.0
+
+### Fixes
+
+- Sentry.framework does not support the minimum OS Version specified in the Info.plist (#3774)
+- Add reference to Swift classes for hybrid SDKs (#3771)
 
 ## 8.22.2